Split invoke tests in 2 files
