+++ .github/workflows/run-appium-smoke-tests-android.yml
@@ -388,7 +388,7 @@ jobs:
       test_suite_tag: SmokeMultiChainAPI
       split_number: ${{ matrix.split }}
       total_splits: 3
-      test-timeout-minutes: 15
+      test-timeout-minutes: 25
       build_type: ${{ inputs.build_type }}
       metamask_environment: ${{ inputs.metamask_environment }}
       runner_provider: ${{ inputs.runner_provider }}
