+++ .github/workflows/run-appium-smoke-tests-ios.yml
@@ -342,7 +342,7 @@ jobs:
       test_suite_tag: SmokeMultiChainAPI
       split_number: ${{ matrix.split }}
       total_splits: 3
-      test-timeout-minutes: 15
+      test-timeout-minutes: 25
       build_type: ${{ inputs.build_type || 'main' }}
       metamask_environment: ${{ inputs.metamask_environment || 'e2e' }}
       runner_provider: ${{ inputs.runner_provider || 'current' }}
